requirements.txt
transformers
sentence-transformers
scikit-learn
pandas
numpy
matplotlib
seaborn
datasets
jupyter
umap-learn
plotly
streamlit
nbformat
